refactoring, moving functions to proper helpers and ratings files
